productivity-agents/IdeationAgnet/requirements.txt
# Python packages for python 3.12
streamlit==1.31.0
requests==2.31.0
python-dotenv==1.0.1
Pillow==10.1.0
openai==1.11.0
pandas==2.1.3
fpdf==1.7.2
langchain==0.1.0
httpx==0.27.2
